requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
absl-py=0.12.0=pypi_0
aenum=3.1.11=pypi_0
aiohttp=3.7.4.post0=pypi_0
async-timeout=3.0.1=pypi_0
attrs=21.2.0=pypi_0
blas=2.116=mkl
blas-devel=3.9.0=16_linux64_mkl
brotlipy=0.7.0=py310h5764c6d_1005
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.12.7=ha878542_0
cachetools=4.2.2=pypi_0
certifi=2020.12.5=pypi_0
cffi=1.15.1=py310h255011f_2
chardet=4.0.0=pypi_0
charset-normalizer=2.0.12=pypi_0
click=8.1.3=pypi_0
colorful=0.5.4=pypi_0
colorlog=6.6.0=pypi_0
cryptography=38.0.3=py310h600f1e7_0
cuda=11.7.1=0
cuda-cccl=11.7.91=0
cuda-command-line-tools=11.7.1=0
cuda-compiler=11.7.1=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cuobjdump=11.7.91=0
cuda-cupti=11.7.101=0
cuda-cuxxfilt=11.7.91=0
cuda-demo-suite=11.8.86=0
cuda-documentation=11.8.86=0
cuda-driver-dev=11.7.99=0
cuda-gdb=11.8.86=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-memcheck=11.8.86=0
cuda-nsight=11.8.86=0
cuda-nsight-compute=11.8.0=0
cuda-nvcc=11.7.99=0
cuda-nvdisasm=11.8.86=0
cuda-nvml-dev=11.7.91=0
cuda-nvprof=11.8.87=0
cuda-nvprune=11.7.91=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-nvvp=11.8.87=0
cuda-runtime=11.7.1=0
cuda-sanitizer-api=11.8.86=0
cuda-toolkit=11.7.1=0
cuda-tools=11.7.1=0
cuda-visual-tools=11.7.1=0
cudatoolkit=11.3.1=ha36c431_9
cython=0.29.32=pypi_0
dataclass-csv=1.3.0=pypi_0
dataclasses-json=0.5.4=pypi_0
docker-pycreds=0.4.0=pypi_0
ffmpeg=4.3=hf484d3e_0
freetype=2.10.4=h0708190_1
fsspec=2021.7.0=pypi_0
future=0.18.2=pypi_0
gds-tools=1.4.0.31=0
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.37.1=pypi_0
hdbscan=0.8.28=pypi_0
idna=2.10=pypi_0
iniconfig=1.1.1=pypi_0
iteration-utilities=0.11.0=pypi_0
jinja2=3.1.1=pypi_0
joblib=1.0.1=pypi_0
jpeg=9e=h166bdaf_2
kubeflow-pytorchjob=0.1.3=pypi_0
kubernetes=17.17.0=pypi_0
lame=3.100=h166bdaf_1003
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=3.0=h9c3ff4c_0
libblas=3.9.0=16_linux64_mkl
libcblas=3.9.0=16_linux64_mkl
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcufile=1.4.0.31=0
libcufile-dev=1.4.0.31=0
libcurand=10.3.0.86=0
libcurand-dev=10.3.0.86=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.10=h7f98852_0
libffi=3.4.2=h7f98852_5
libgcc-ng=12.1.0=h8d9b700_16
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libgomp=12.1.0=h8d9b700_16
libiconv=1.17=h166bdaf_0
liblapack=3.9.0=16_linux64_mkl
liblapacke=3.9.0=16_linux64_mkl
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnsl=2.0.0=h7f98852_0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.37=h21135ba_2
libstdcxx-ng=12.2.0=h46fd767_19
libtiff=4.3.0=h0fcbabc_4
libuuid=2.32.1=h7f98852_1000
libwebp-base=1.2.4=h166bdaf_0
libzlib=1.2.11=h166bdaf_1014
llvm-openmp=14.0.4=he0ac6c6_0
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.3.4=pypi_0
markupsafe=2.1.1=pypi_0
marshmallow=3.13.0=pypi_0
marshmallow-enum=1.5.1=pypi_0
mkl=2022.1.0=h84fe81f_915
mkl-devel=2022.1.0=ha770c72_916
mkl-include=2022.1.0=h84fe81f_915
multidict=5.1.0=pypi_0
mypy-extensions=0.4.3=pypi_0
ncurses=6.3=h27087fc_1
nettle=3.6=he412f7d_0
nsight-compute=2022.3.0.22=0
numpy=1.22.3=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
oauthlib=3.2.0=pypi_0
olefile=0.46=pyh9f0ad1d_1
openh264=2.1.1=h780b84a_0
openjpeg=2.5.0=h7d73246_0
openssl=3.0.7=h0b41bf4_1
packaging=21.0=pypi_0
parameterized=0.8.1=pypi_0
pathtools=0.1.2=pypi_0
pillow=9.3.0=pypi_0
pint=0.17=pypi_0
pip=22.0.4=pyhd8ed1ab_0
pluggy=1.0.0=pypi_0
prettyprinter=0.18.0=pypi_0
promise=2.3=pypi_0
protobuf=3.16.0=pypi_0
psutil=5.8.0=pypi_0
py=1.11.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pydeprecate=0.3.1=pypi_0
pygments=2.10.0=pypi_0
pyopenssl=22.1.0=pyhd8ed1ab_0
pyparsing=2.4.7=pypi_0
pysocks=1.7.1=pyha2e5f31_6
pytest=7.1.2=pypi_0
python=3.10.4=h2660328_0_cpython
python-dateutil=2.8.2=pypi_0
python-dotenv=0.17.1=pypi_0
python_abi=3.10=2_cp310
pytorch=1.13.0=py3.10_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h67b0de4_0
pytorch-mutex=1.0=cuda
pytz=2021.1=pypi_0
pyyaml=5.4.1=pypi_0
readline=8.1=h46c0cb4_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
retrying=1.3.3=pypi_0
rsa=4.7.2=pypi_0
schedule=1.1.0=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.8.0=pypi_0
sentry-sdk=1.5.12=pypi_0
setproctitle=1.2.3=pypi_0
setuptools=62.3.1=py310hff52083_0
shortuuid=1.0.9=pypi_0
six=1.16.0=pypi_0
smmap=5.0.0=pypi_0
sqlite=3.38.5=h4ff8645_0
stringcase=1.2.0=pypi_0
table-logger=0.3.6=pypi_0
tbb=2021.6.0=h924138e_1
tensorboard=2.5.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.12=h27826a3_0
tomli=2.0.1=pypi_0
torchaudio=0.13.0=py310_cu117
torchmetrics=0.5.0=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.14.0=py310_cu117
tqdm=4.49.0=pypi_0
typing-extensions=4.1.1=pypi_0
typing-inspect=0.7.1=pypi_0
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022a=h191b570_0
urllib3=1.26.5=pypi_0
wandb=0.12.16=pypi_0
websocket-client=1.2.0=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.37.1=pyhd8ed1ab_0
xz=5.2.5=h516909a_1
yarl=1.6.3=pypi_0
zlib=1.2.11=h166bdaf_1014
zstd=1.5.2=h8a70e8d_1
